requirements/common.txt
boto3==1.13.19
botocore==1.16.19
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-allauth==0.41.0
django-countries==6.1.2
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filebrowser-no-grappelli==3.8.0
django-heroku==0.3.1
django-storages==1.7.1
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.10.0
jsmin==2.2.2
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
Pillow==7.0.0
psycopg2==2.7.5
psycopg2-binary==2.8.5
pydot==1.4.1
pygraphviz==1.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
stripe==2.47.0
urllib3==1.25.8
whitenoise==5.1.0
